refactor(signals): extract signal utilities for cleaner event handling
