+++ src/signals/signal_utils.c
@@ -1,0 +1,61 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal_utils.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: thanh-ng <thanh-ng@student.42vienna.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/03/08 14:00:00 by thanh-ng          #+#    #+#             */
+/*   Updated: 2026/03/08 14:00:00 by thanh-ng         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	handle_child_exit(int *last_exit_status, pid_t pid)
+{
+	int	status;
+
+	status = 0;
+	if (waitpid(pid, &status, 0) == -1)
+	{
+		*last_exit_status = 1;
+		return (-1);
+	}
+	if (WIFSIGNALED(status))
+	{
+		*last_exit_status = 128 + WTERMSIG(status);
+		if (WTERMSIG(status) == SIGQUIT)
+			ft_putstr_fd("Quit (core dumped)\n", STDOUT_FILENO);
+		else if (WTERMSIG(status) == SIGINT)
+			ft_putstr_fd("\n", STDOUT_FILENO);
+	}
+	else if (WIFEXITED(status))
+		*last_exit_status = WEXITSTATUS(status);
+	else
+		*last_exit_status = EXIT_FAILURE;
+	return (0);
+}
+
+int	readline_event_hook(void)
+{
+	if (g_signum == SIGINT)
+	{
+		g_signum = 0;
+		rl_on_new_line();
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}
+	return (0);
+}
+
+int	check_signal_received(t_shell *shell)
+{
+	if (g_signum == SIGINT)
+	{
+		shell->last_exit = 130;
+		g_signum = 0;
+		return (1);
+	}
+	return (0);
+}
